Updated exports and renamed fields to field
