+++ package/src/lib/overlay/index.ts
@@ -2,4 +2,5 @@ import Screen from "./screen/Screen.svelte";
 import Editor from "./editor/Editor.svelte";
 
 export { Screen, Editor };
+export { field } from "./field/index.js";
 export type { WebComponent } from "./source/types.js";
